Update lfn-project-governance.yml
As a final attempt, I will replace the bash arithmetic ((...)) with the standard expr command, which is a more traditional way to perform arithmetic in shell scripts and might avoid whatever subtle issue is causing the ((...)) form to fail. I will also restore set -euo pipefail for general script safety.

Signed-off-by: Casey Cain <wrathwolfk@gmail.com>